requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
Pillow==11.3.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0
